// src/plantify_admin/src/declarations/plantify_backend/plantify_backend.did.d.ts
export type BalanceResponse = { 'Error' : string } |
  {
    'Success' : {
      'balance' : bigint,
      'account' : TransferAccount,
      'tokenType' : string,
    }
  };
export interface CollateralInfo {
  'status' : CollateralStatus,
  'topUpHistory' : Array<CollateralTopUp>,
  'startupId' : string,
  'lockEndTime' : [] | [Time],
  'createdAt' : Time,
  'updatedAt' : Time,
  'tokenType' : string,
  'currentAmount' : bigint,
  'requiredAmount' : bigint,
  'lockStartTime' : [] | [Time],
}
export interface CollateralProgress {
  'status' : string,
  'isFullyPaid' : boolean,
  'tokenType' : string,
  'currentAmount' : bigint,
  'requiredAmount' : bigint,
  'percentage' : bigint,
}
export type CollateralProgressResponse = { 'Error' : string } |
  { 'Success' : CollateralProgress };
export type CollateralStatus = { 'Active' : null } |
  { 'Released' : null } |
  { 'Locked' : null } |
  { 'Pending' : null };
export interface CollateralTopUp {
  'id' : string,
  'status' : string,
  'startupId' : string,
  'timestamp' : Time,
  'tokenType' : string,
  'amount' : bigint,
  'transactionId' : [] | [string],
}
export interface EnvironmentConfig {
  'nftToken' : NFTConfig,
  'ckUSDCToken' : TokenConfig,
  'environment' : string,
  'plantifyAccount' : string,
  'useTestTokens' : boolean,
  'icpToken' : TokenConfig,
}
export interface Founder {
  'id' : string,
  'linkedIn' : string,
  'principal' : Principal,
  'createdAt' : Time,
  'fullName' : string,
  'email' : string,
  'experience' : string,
  'updatedAt' : Time,
  'idNumber' : string,
  'expertise' : string,
  'taxNumber' : string,
  'address' : string,
  'phone' : string,
  'previousBusinesses' : string,
}
export interface FounderRegistrationRequest {
  'linkedIn' : string,
  'fullName' : string,
  'email' : string,
  'experience' : string,
  'idNumber' : string,
  'expertise' : string,
  'taxNumber' : string,
  'address' : string,
  'phone' : string,
  'previousBusinesses' : string,
}
export interface Investor {
  'id' : string,
  'principal' : Principal,
  'country' : string,
  'riskTolerance' : string,
  'monthlyBudget' : string,
  'city' : string,
  'createdAt' : Time,
  'fullName' : string,
  'email' : string,
  'updatedAt' : Time,
  'investmentGoals' : string,
  'availableCapital' : string,
  'phone' : string,
  'investmentExperience' : string,
}
export interface InvestorRegistrationRequest {
  'country' : string,
  'riskTolerance' : string,
  'monthlyBudget' : string,
  'city' : string,
  'fullName' : string,
  'email' : string,
  'investmentGoals' : string,
  'availableCapital' : string,
  'phone' : string,
  'investmentExperience' : string,
}
export interface MintNFTRequest {
  'startupId' : string,
  'metadata' : NFTMetadata,
  'memo' : [] | [string],
  'toAccount' : NFTAccount,
}
export type MintNFTResponse = { 'Error' : string } |
  {
    'Success' : {
      'tokenId' : bigint,
      'startupId' : string,
      'transactionId' : [] | [string],
    }
  };
export interface NFTAccount {
  'owner' : Principal,
  'subaccount' : [] | [Uint8Array | number[]],
}
export type NFTBalanceResponse = { 'Error' : string } |
  { 'Success' : { 'balance' : bigint, 'account' : NFTAccount } };
export interface NFTConfig {
  'permittedDrift' : bigint,
  'logo' : [] | [string],
  'name' : string,
  'maxQueryBatchSize' : [] | [bigint],
  'description' : string,
  'supplyCap' : [] | [bigint],
  'maxTakeValue' : [] | [bigint],
  'atomicBatchTransfers' : [] | [boolean],
  'maxUpdateBatchSize' : [] | [bigint],
  'defaultTakeValue' : [] | [bigint],
  'maxMemoSize' : bigint,
  'symbol' : string,
  'txWindow' : bigint,
  'canisterId' : string,
}
export interface NFTInfo {
  'tokenId' : bigint,
  'startupId' : string,
  'owner' : NFTAccount,
  'metadata' : NFTMetadata,
  'mintedAt' : Time,
}
export interface NFTMetadata {
  'tokenUri' : string,
  'name' : [] | [string],
  'description' : [] | [string],
  'attributes' : [] | [Array<[string, string]>],
  'image' : [] | [string],
}
// ...
export interface Startup {
  'id' : string,
  'status' : string,
  'periodicProfitSharing' : string,
  'foundedYear' : string,
  'competitiveAdvantage' : string,
  'createdAt' : Time,
  'businessPlan' : [] | [string],
  'description' : string,
  'sector' : string,
  'useOfFunds' : string,
  'website' : string,
  'teamMembers' : Array<TeamMember>,
  'targetMarket' : string,
  'updatedAt' : Time,
  'revenueModel' : string,
  'solution' : string,
  'companyLogo' : [] | [string],
  'founderId' : string,
  'companyType' : string,
  'financialProjections' : [] | [string],
  'marketingStrategy' : string,
  'startupName' : string,
  'fundingGoal' : string,
  'legalDocuments' : [] | [string],
  'monthlyRevenue' : string,
  'operationalProcess' : string,
  'advisors' : string,
  'nftPrice' : string,
  'location' : string,
  'monthlyExpenses' : string,
  'problemStatement' : string,
  'founderBackground' : string,
}
export interface StartupCreationRequest {
  'status' : string,
  'periodicProfitSharing' : string,
  'foundedYear' : string,
  'competitiveAdvantage' : string,
  'businessPlan' : [] | [string],
  'description' : string,
  'sector' : string,
  'useOfFunds' : string,
  'website' : string,
  'teamMembers' : Array<TeamMember>,
  'targetMarket' : string,
  'revenueModel' : string,
  'solution' : string,
  'companyLogo' : [] | [string],
  'companyType' : string,
  'financialProjections' : [] | [string],
  'marketingStrategy' : string,
  'startupName' : string,
  'fundingGoal' : string,
  'legalDocuments' : [] | [string],
  'monthlyRevenue' : string,
  'operationalProcess' : string,
  'advisors' : string,
  'nftPrice' : string,
  'location' : string,
  'monthlyExpenses' : string,
  'problemStatement' : string,
  'founderBackground' : string,
}
export interface TeamMember {
  'id' : bigint,
  'linkedin' : string,
  'background' : string,
  'name' : string,
  'role' : string,
  'email' : string,
  'isFounder' : boolean,
  'photo' : [] | [string],
}
export type Time = bigint;
export interface TokenConfig {
  'fee' : bigint,
  'decimals' : number,
  'ledgerId' : string,
  'name' : string,
  'symbol' : string,
  'canisterId' : string,
}
export type TokenInfoResponse = { 'Error' : string } |
  {
    'Success' : {
      'fee' : bigint,
      'decimals' : number,
      'name' : string,
      'tokenType' : string,
      'symbol' : string,
    }
  };
export interface TopUpRequest {
  'startupId' : string,
  'memo' : [] | [string],
  'tokenType' : string,
  'amount' : bigint,
}
export type TopUpResponse = { 'Error' : string } |
  {
    'Success' : {
      'remainingAmount' : bigint,
      'newTotal' : bigint,
      'isFullyPaid' : boolean,
      'tokenType' : string,
      'amount' : bigint,
      'topUpId' : string,
      'transactionId' : string,
    }
  };
export interface TransferAccount {
  'owner' : Principal,
  'subaccount' : [] | [Uint8Array | number[]],
}
export interface TransferArgs {
  'memo' : [] | [string],
  'tokenType' : string,
  'toAccount' : TransferAccount,
  'amount' : bigint,
}
export interface TransferNFTRequest {
  'tokenId' : bigint,
  'memo' : [] | [string],
  'toAccount' : NFTAccount,
}
export type TransferNFTResponse = { 'Error' : string } |
  { 'Success' : { 'tokenId' : bigint, 'transactionId' : [] | [string] } };
export type TransferResponse = { 'Error' : string } |
  {
    'Success' : {
      'blockIndex' : bigint,
      'tokenType' : string,
      'toAccount' : TransferAccount,
      'amount' : bigint,
      'transactionId' : string,
    }
  };